+++ backend/src/main/java/cn/edu/shiep/backend/meetingroom/dto/request/ReservationRequest.java
@@ -1,6 +1,8 @@
 package cn.edu.shiep.backend.meetingroom.dto.request;
 
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -8,7 +10,7 @@ import java.util.List;
 
 @Data
 public class ReservationRequest {
-    private Integer roomId;
+    private Long roomId;
     private String theme;
     private int personNum;
     private LocalDateTime startTime;
