+++ backend/src/main/java/cn/edu/shiep/backend/meetingroom/repository/ConferenceRoomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 // 数据访问接口
 @Repository
-public interface ConferenceRoomRepository extends JpaRepository<ConferenceRoom, Integer> {
+public interface ConferenceRoomRepository extends JpaRepository<ConferenceRoom, Long> {
 }
