+++ backend/src/main/java/cn/edu/shiep/backend/meetingroom/repository/RoleRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Roles, Integer> {
+public interface RoleRepository extends JpaRepository<Roles, Long> {
     Optional<Roles> findByName(ERole name);
 }
